+++ packages/core/src/languages/css/index.ts
@@ -62,7 +62,7 @@ export type CssLanguagePluginOptions = {
  *   ],
  * });
  * @since v0.1.0
- * @version v0.1.21
+ * @version v0.1.22
  */
 export default class CssLanguagePlugin extends SimpleLanguagePlugin {
   /**